Medical is mostly up (plus other changes)
Added a lot of content to medical and the possibility of expanding it at a later date.
Also added in a cat, a lizard plushie and an otie to certain council rooms (upper floor), plus a somewhat barebones but functional gym north of the spa and a WIP shower room connected to said gym. (lower floor)