Cleaning Ckey and Custom things
Removed a lot of defines for custom clothes/items and stuff
We can look through sprites and readd whatever we want later without the special name-referencing descriptions and stuff.
Commented out the checks for ckey on sprite_accessories, might break things so left the code in there just commented out and removed ckey assignments entirely off of the accessories.