Paints! And more small tweaks.
Added easy paint effects for walls, for both the wall colors and the stripes!
Tried to rework some things to use Color variables instead of a hard coded mess.
Refactored the Prosthetics Machine in robotics to be the proper one.
Left an example of the paint use on Stargazer deck 1.
Figured out that active edges for atmos are the cause of the crashing issues. For now, it is best to leave it compiling Stargazer whenever you push. But expect it to crash on trying to close if you load the Tether to test things on a functioning, full Station.